create table exercise

+++ albums_migration.sql
@@ -7,7 +7,7 @@ CREATE TABLE albums (
     artist VARCHAR(50) NOT NULL,
     song VARCHAR(50) NOT NULL,
     release_date INT UNSIGNED NOT NULL,
-    sales DECIMAL(4,2) UNSIGNED,
+    sales FLOAT UNSIGNED,
     genre VARCHAR(50) NOT NULL,
     PRIMARY KEY (id)
 );
